requirements.txt
# Core dependencies
requests>=2.31.0
httpx>=0.25.0
pyyaml>=6.0
python-dotenv>=1.0.0

# Data & Database
sqlalchemy>=2.0.0
pandas>=2.1.0

# User Agents & Fingerprinting
fake-useragent>=1.4.0
user-agents>=2.2.0

# CLI Dashboard
rich>=13.7.0
textual>=0.47.0

# Web Dashboard (Optional)
fastapi>=0.109.0
uvicorn>=0.27.0
jinja2>=3.1.3

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Utilities
tqdm>=4.66.0
loguru>=0.7.0